# packages/items_service_client/docs/openapi.yaml
components:
    examples: {}
    headers: {}
    parameters: {}
    requestBodies: {}
    responses: {}
    schemas:
        ItemStatus:
            type: string
            enum:
                - Queued
            nullable: false
        LocationProps:
            properties:
                positions:
                    properties:
                        begin:
                            properties:
                                line:
                                    type: number
                                    format: double
                            required:
                                - line
                            type: object
                    required:
                        - begin
                    type: object
                path:
                    type: string
            required:
                - positions
                - path
            type: object
        Severity:
            type: string
            enum:
                - HIGH
            nullable: false
        MetadataProps:
            properties:
                severity:
                    $ref: '#/components/schemas/Severity'
                description:
                    type: string
            required:
                - severity
                - description
            type: object
        FindingProps:
            properties:
                id:
                    type: number
                    format: double
                itemId:
                    type: number
                    format: double
                type:
                    type: string
                ruleId:
                    type: string
                location:
                    $ref: '#/components/schemas/LocationProps'
                metadata:
                    $ref: '#/components/schemas/MetadataProps'
            required:
                - id
                - itemId
                - type
                - ruleId
                - location
                - metadata
            type: object
            additionalProperties: false
        ItemProps:
            properties:
                id:
                    type: number
                    format: double
                status:
                    $ref: '#/components/schemas/ItemStatus'
                repositoryName:
                    type: string
                findings:
                    items:
                        $ref: '#/components/schemas/FindingProps'
                    type: array
                queuedAt:
                    type: string
                scanningAt:
                    type: string
                finishedAt:
                    type: string
            required:
                - id
                - status
                - repositoryName
                - findings
                - queuedAt
            type: object
            additionalProperties: false
        Pick_ItemProps.Exclude_keyofItemProps.id-or-findings__:
            properties: {}
            type: object
            description: 'From T, pick a set of properties whose keys are in the union K'
        Omit_ItemProps.id-or-findings_:
            $ref: '#/components/schemas/Pick_ItemProps.Exclude_keyofItemProps.id-or-findings__'
        Partial_Pick_ItemProps.id-or-findings__:
            properties: {}
            type: object
            description: 'Make all properties in T optional'
        PartialBy_ItemProps.id-or-findings_:
            allOf:
                -
                    $ref: '#/components/schemas/Omit_ItemProps.id-or-findings_'
                -
                    $ref: '#/components/schemas/Partial_Pick_ItemProps.id-or-findings__'
        Pick_FindingProps.Exclude_keyofFindingProps.id__:
            properties: {}
            type: object
            description: 'From T, pick a set of properties whose keys are in the union K'
        Omit_FindingProps.id_:
            $ref: '#/components/schemas/Pick_FindingProps.Exclude_keyofFindingProps.id__'
        Partial_Pick_FindingProps.id__:
            properties: {}
            type: object
            description: 'Make all properties in T optional'
        PartialBy_FindingProps.id_:
            allOf:
                -
                    $ref: '#/components/schemas/Omit_FindingProps.id_'
                -
                    $ref: '#/components/schemas/Partial_Pick_FindingProps.id__'
        FindingCreationParams:
            $ref: '#/components/schemas/PartialBy_FindingProps.id_'
        ItemCreationParams:
            allOf:
                -
                    $ref: '#/components/schemas/PartialBy_ItemProps.id-or-findings_'
                -
                    properties:
                        findings:
                            items:
                                $ref: '#/components/schemas/FindingCreationParams'
                            type: array
                    type: object
    securitySchemes: {}
info:
    title: items-service
    version: 1.0.0
    license:
        name: ISC
    contact: {}
openapi: 3.0.0
paths:
    /healthcheck:
        get:
            operationId: HealthCheck
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    status: {type: string}
                                required:
                                    - status
                                type: object
            security: []
            parameters: []
    /api/v1/items:
        get:
            operationId: GetItems
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    data: {items: {$ref: '#/components/schemas/ItemProps'}, type: array}
                                required:
                                    - data
                                type: object
            security: []
            parameters: []
    '/api/v1/items/{itemId}':
        get:
            operationId: GetItem
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    data: {$ref: '#/components/schemas/ItemProps'}
                                required:
                                    - data
                                type: object
            security: []
            parameters:
                -
                    in: path
                    name: itemId
                    required: true
                    schema:
                        type: string
    /api/v1/item:
        post:
            operationId: CreateItem
            responses:
                '201':
                    description: Created
                    content:
                        application/json:
                            schema:
                                properties:
                                    data: {$ref: '#/components/schemas/ItemProps'}
                                required:
                                    - data
                                type: object
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ItemCreationParams'
servers:
    -
        url: /