# packages/items_service_client/src/model/findingCreationParams.ts
/**
 * Generated by orval v6.8.1 🍺
 * Do not edit manually.
 * items-service
 * OpenAPI spec version: 1.0.0
 */
import type { PartialByFindingPropsId } from './partialByFindingPropsId';

export type FindingCreationParams = PartialByFindingPropsId;