# packages/items_service_client/src/model/findingProps.ts
/**
 * Generated by orval v6.8.1 🍺
 * Do not edit manually.
 * items-service
 * OpenAPI spec version: 1.0.0
 */
import type { LocationProps } from './locationProps';
import type { MetadataProps } from './metadataProps';

export interface FindingProps {
  id: number;
  itemId: number;
  type: string;
  ruleId: string;
  location: LocationProps;
  metadata: MetadataProps;
}