# packages/items_service_client/src/model/itemCreationParams.ts
/**
 * Generated by orval v6.8.1 🍺
 * Do not edit manually.
 * items-service
 * OpenAPI spec version: 1.0.0
 */
import type { PartialByItemPropsIdOrFindings } from './partialByItemPropsIdOrFindings';
import type { ItemCreationParamsAllOf } from './itemCreationParamsAllOf';

export type ItemCreationParams = PartialByItemPropsIdOrFindings & ItemCreationParamsAllOf;