# packages/items_service_client/src/model/itemProps.ts
/**
 * Generated by orval v6.8.1 🍺
 * Do not edit manually.
 * items-service
 * OpenAPI spec version: 1.0.0
 */
import type { ItemStatus } from './itemStatus';
import type { FindingProps } from './findingProps';

export interface ItemProps {
  id: number;
  status: ItemStatus;
  repositoryName: string;
  findings: FindingProps[];
  queuedAt: string;
  scanningAt?: string;
  finishedAt?: string;
}